Tell compiler we know what we are doing.

+++ g_utils.c
@@ -1946,7 +1946,7 @@ void highlight_text(char* src, char* dest)
 {
 	if (dest == NULL)
 		dest = src;
-	convert_string(src, 0, 0x7f, 0x80, dest); // white -> green
+	convert_string(src, 0, (byte)0x7f, (byte)0x80, dest); // white -> green
 }
 
 /**
@@ -1959,7 +1959,7 @@ void white_text(char* src, char* dest)
 {
 	if (dest == NULL)
 		dest = src;
-	convert_string(src, 0x80, 0xff, -128, dest); // green -> white
+	convert_string(src, (byte)0x80, (byte)0xff, -128, dest); // green -> white
 }
 
 /**
